README
Very important edit:

The full stop (Commonwealth English), period (North American English) or full point . is a punctuation mark. It is used for several purposes, most often to mark the end of a declarative sentence (as opposed to a question or exclamation); this sentence-terminal use, alone, defines the strictest sense of full stop.

The mark is also often used, singly, to indicate omitted characters or in a series, as an ellipsis (…), to indicate omitted words. It may be placed after an initial letter used to stand for a name, or after each individual letter in an initialism or acronym (e.g., "U.S.A."); however, this style is declining, and many initialisms without punctuation (e.g., "UK" and "NATO") have become accepted norms. A full point is also frequently used at the end of word abbreviations – in British usage, primarily truncations like Rev., but not after contractions like Revd (in American English it is used in both cases).

In Anglophone countries, it is used for the decimal point and other purposes, and may be called a point. In computing, it is called a dot.[1] It is sometimes called a baseline dot to distinguish it from the interpunct (or middle dot).[1][2] While full stop technically only applies to the full point when used to terminate a sentence, the distinction – drawn since at least 1897[3] – is not maintained by all modern style guides and dictionaries.