arena: Replace excesive '>=' with '>'
Imagine the following arena:

             a.size
           <-------->
        ###.....     ###
    a.data ^    ^ a.offset

Where '#' represents unusable space, '.' means allocated chunks, and ' ' means
unallocated chunks in the arena. The current offset is 5, which is both the
number of allocated bytes, and the index for the next allocation.

If the user tries to allocate another 5 bytes, the offset will grow to 10, which
is equal to the arena size. This offset will be incorrect for future
allocations, but not currently, since the 10 bytes fit in the arena.